Extract validator boolean expression helpers
